Factored out MatchInfo from MatchModel.
