+++ paq8px.vcxproj
@@ -299,6 +299,7 @@
     <ClCompile Include="model\Image4BitModel.cpp" />
     <ClCompile Include="model\Image8BitModel.cpp" />
     <ClCompile Include="model\IndirectModel.cpp" />
+    <ClCompile Include="model\MatchInfo.cpp" />
     <ClCompile Include="model\WordModelInfo.cpp" />
     <ClCompile Include="model\JpegModel.cpp" />
     <ClCompile Include="model\LinearPredictionModel.cpp" />
@@ -444,6 +445,7 @@
     <ClInclude Include="model\Image8BitModel.hpp" />
     <ClInclude Include="model\ImageModelsCommon.hpp" />
     <ClInclude Include="model\IndirectModel.hpp" />
+    <ClInclude Include="model\MatchInfo.hpp" />
     <ClInclude Include="model\WordModelInfo.hpp" />
     <ClInclude Include="model\JpegModel.hpp" />
     <ClInclude Include="model\LinearPredictionModel.hpp" />
